Allow assignment of picker hitter/pitcher disposition
The currently generated simulated teams are as a whole league picking
the correct players, but every team making their choice with the same
ranking logic is resulting in counting stats for offense having standard
deviations that average about 20% lower than the standard deviations we
observed in our league in actual play (the means are within 5% in every
stat).

This change allows the assignment of dispositions to a picker. If a
picker has a disposition of 1.0, they pick the top player that fits
their roster by the consensus value. If the disposition is higher than
1.0, the team will "overvalue" hitters by re-sorting the candidates and
giving hitters a bump. If the disposition is below one, hitters get a
penalty.